Update the framework.
